day9 cleanup + intcode improvements
I spent more time than I'd like trying to work around the limitations
inherent in the Index trait (while trying to make it silently grow
memory if an uninitialized location was accesed). In the end, I realized
that there's literally no way to get the Index syntax to work how I
wanted (it would need to return a concrete value, not a reference), so I
caved and rewrote everything to call mem.read and mem.write.

That said, I did do some shenanigans with RefCells to keep Mem::read as
a &self. Why? Well, some might argue that since reads are semantically
immutable operations (in the intcode machine), it would make sense to
have the function signature reflect that. That's fair, but to tell you
the truth, the reason _I_ did it was so that I didn't have to rewrite my
janky lil inline closure for fetching arguments haha.

---

I also cleaned up the instruction decoding somewhat, pushing it back
into the main cpu.rs file (mainly for modification speed, though also to
avoig having to pass CPU state to the decode function redundantly), and
tweaking the enum args to reflect the actual type of argument they
represent.