reduce memory footprint when accessing large addrs
Resizing the vector when accessing large values is a _baaaad_ idea. If
you have a buggy intcode program, and you try and access a massive
address for some reason, the vec will baloon in size for no good reason!

This new approach splits the memory into "low memory" and "high memory",
backed by a Vec and a HashMap respectively.